fix Step 1 리뷰내용 적용 2020-06-26 18:59

+++ src/main/java/qna/domain/Question.java
@@ -97,7 +97,8 @@ public class Question extends AbstractEntity {
         return deleted;
     }
 
-    public Question delete(List<DeleteHistory> deleteHistories) {
+    public Question delete(List<DeleteHistory> deleteHistories, User loginUser) throws CannotDeleteException {
+        checkOwner(loginUser);
         this.setDeleted(true);
         deleteHistories.add(new DeleteHistory(ContentType.QUESTION, this.getId(), this.getWriter(), LocalDateTime.now()));
         answers.delete(deleteHistories);
